Add purchase-cart logic into item module

+++ spring-cloud-microservice-item/src/main/java/com/zhilong/springcloud/controller/ItemController.java
@@ -53,7 +53,7 @@ public class ItemController {
     }
 
     @GetMapping("/brief/{id}")
-    public ResponseEntity<List<ItemSimpleTo>> itemSimple(@RequestParam(value = "page", defaultValue = "0") Integer page, @RequestParam(value = "size",defaultValue = "4") Integer size, @PathVariable("id") Integer id){
+    public ResponseEntity<List<ItemSimpleTo>> itemSimple(@RequestParam(value = "page", defaultValue = "0") Integer page, @RequestParam(value = "size",defaultValue = "4") Integer size, @PathVariable("id") Long id){
         List<ItemSimpleTo> all = new ArrayList<>();
         all = itemService.findAllSimpleItemByCategory(id,PageRequest.of(page, size));
         logger.info("Finding a part of items with size : " + all.size() + ". Page : " + page);
