+++ spring-cloud-microservice-item/src/main/java/com/zhilong/springcloud/entity/dto/ItemSimpleTo.java
@@ -13,12 +13,12 @@ public class ItemSimpleTo implements Serializable {
     private String image;
     private String item_des;
     private String title;
-    private Integer cat_id;
+    private Long cat_id;
 
     public ItemSimpleTo() {
     }
 
-    public ItemSimpleTo(Long id, BigDecimal currentprice, String image, String item_des, String title, Integer cat_id) {
+    public ItemSimpleTo(Long id, BigDecimal currentprice, String image, String item_des, String title, Long cat_id) {
         this.id = id;
         this.currentprice = currentprice;
         this.image = image;
